+++ ootrando_overworldmap_hamsda/scripts/data_per_region.lua
@@ -1776,7 +1776,7 @@ data_per_region = {
           return AccessibilityLevel.None
         end
       },
-      ["GF North F1 Carpenter"] = {
+      ["Hideout Jail Guard (1 Torch)"] = {
         ["child_access"] = function()
           if has("sword1") then
             return AccessibilityLevel.Normal
@@ -1789,7 +1789,20 @@ data_per_region = {
           return AccessibilityLevel.Normal
         end
       },
-      ["GF North F2 Carpenter"] = {
+      ["Hideout Jail Guard (2 Torches)"] = {
+        ["child_access"] = function()
+          if has("sword1") then
+            return AccessibilityLevel.Normal
+          elseif has("sticks") then
+            return AccessibilityLevel.SequenceBreak
+          end
+          return AccessibilityLevel.None
+        end,
+        ["adult_access"] = function()
+          return AccessibilityLevel.Normal
+        end
+      },
+      ["Hideout Jail Guard (3 Torches)"] = {
         ["child_access"] = function()
           local weapon = AccessibilityLevel.None
           if has("sword1") then
@@ -1810,7 +1823,7 @@ data_per_region = {
           return AccessibilityLevel.SequenceBreak
         end
       },
-      ["GF South F1 Carpenter"] = {
+      ["Hideout Jail Guard (4 Torches)"] = {
         ["child_access"] = function()
           if has("sword1") then
             return AccessibilityLevel.Normal
@@ -1823,20 +1836,7 @@ data_per_region = {
           return AccessibilityLevel.Normal
         end
       },
-      ["GF South F2 Carpenter"] = {
-        ["child_access"] = function()
-          if has("sword1") then
-            return AccessibilityLevel.Normal
-          elseif has("sticks") then
-            return AccessibilityLevel.SequenceBreak
-          end
-          return AccessibilityLevel.None
-        end,
-        ["adult_access"] = function()
-          return AccessibilityLevel.Normal
-        end
-      },
-      ["GF Gerudo Membership Card"] = {
+      ["Hideout Gerudo Membership Card"] = {
         ["child_access"] = function()
           return can_finish_GerudoFortress("child")
         end,
@@ -1893,7 +1893,7 @@ data_per_region = {
           return AccessibilityLevel.None
         end
       },
-      ["Gerudo Training Grounds Split"] = {
+      ["Gerudo Training Ground Split"] = {
         ["setting"] = "setting_entrance_dungeons_shuffle",
         ["child_access"] = function()
           return AccessibilityLevel.None
@@ -3728,7 +3728,7 @@ data_per_region = {
           return AccessibilityLevel.Normal
         end
       },
-      ["Graveyard Composers Grave"] = {
+      ["Royal Familys Tomb"] = {
         ["setting"] = "setting_entrance_grottos_shuffle",
         ["child_access"] = function()
           if has("ocarina") and has("lullaby") then
@@ -3835,11 +3835,11 @@ data_per_region = {
       }
     }
   },
-  ["Graveyard Composers Grave"] = {
-    ["scene"] = "Graveyard Composers Grave",
+  ["Royal Familys Tomb"] = {
+    ["scene"] = "Royal Familys Tomb",
     ["interior"] = true,
     ["locations"] = {
-      ["Graveyard Composers Grave Chest"] = {
+      ["Royal Familys Tomb Chest"] = {
         ["child_access"] = function()
           if has("dinsfire") and has("magic") then
             return AccessibilityLevel.Normal
@@ -3852,7 +3852,7 @@ data_per_region = {
           return has_fire()
         end
       },
-      ["Song from Composers Grave"] = {
+      ["Song from Royal Familys Tomb"] = {
         ["child_access"] = function()
           if has("sling") or has("boomerang") or has("sticks") or has("sword1") then
             return AccessibilityLevel.Normal
@@ -11198,11 +11198,11 @@ data_per_region = {
       }
     }
   },
-  ["Gerudo Training Grounds Split"] = {
-    ["scene"] = "Gerudo Training Grounds",
+  ["Gerudo Training Ground Split"] = {
+    ["scene"] = "Gerudo Training Ground",
     ["dungeon"] = true,
     ["exits"] = {
-      ["Gerudo Training Grounds Lobby"] = {
+      ["Gerudo Training Ground Lobby"] = {
         ["child_access"] = function()
           if has("gtg_reg") then
             return AccessibilityLevel.Normal
@@ -11216,7 +11216,7 @@ data_per_region = {
           return AccessibilityLevel.None
         end
       },
-      ["Gerudo Training Grounds MQ Lobby"] = {
+      ["Gerudo Training Ground MQ Lobby"] = {
         ["child_access"] = function()
           if has("gtg_mq") then
             return AccessibilityLevel.Normal
@@ -11232,11 +11232,11 @@ data_per_region = {
       }
     }
   },
-  ["Gerudo Training Grounds Lobby"] = {
-    ["scene"] = "Gerudo Training Grounds",
+  ["Gerudo Training Ground Lobby"] = {
+    ["scene"] = "Gerudo Training Ground",
     ["dungeon"] = true,
     ["locations"] = {
-      ["Gerudo Training Grounds Lobby Chests"] = {
+      ["Gerudo Training Ground Lobby Chests"] = {
         ["child_access"] = function()
           if has("sling") then
             return AccessibilityLevel.Normal
@@ -11250,7 +11250,7 @@ data_per_region = {
           return AccessibilityLevel.None
         end
       },
-      ["Gerudo Training Grounds Stalfos Chest"] = {
+      ["Gerudo Training Ground Stalfos Chest"] = {
         ["child_access"] = function()
           if has("sword1") then
             return AccessibilityLevel.Normal
@@ -11263,7 +11263,7 @@ data_per_region = {
           return AccessibilityLevel.Normal
         end
       },
-      ["Gerudo Training Grounds Beamos Chest"] = {
+      ["Gerudo Training Ground Beamos Chest"] = {
         ["child_access"] = function()
           local weapon = AccessibilityLevel.None
           if has("sword1") then
@@ -11287,7 +11287,7 @@ data_per_region = {
           return AccessibilityLevel.Normal
         end
       },
-      ["Gerudo Training Grounds Heavy Block Room"] = {
+      ["Gerudo Training Ground Heavy Block Room"] = {
         ["child_access"] = function()
           if has("sword1") and has("logic_gtg_without_hookshot") then
             return AccessibilityLevel.Normal
@@ -11303,10 +11303,10 @@ data_per_region = {
           return AccessibilityLevel.SequenceBreak
         end
       },
-      ["Gerudo Training Grounds Lava Room"] = {
+      ["Gerudo Training Ground Lava Room"] = {
         ["child_access"] = function()
           local explo = has_explosives()
-          if has("sword1") or (access_region("Gerudo Training Grounds Lobby", "adult") == AccessibilityLevel.Normal) then
+          if has("sword1") or (access_region("Gerudo Training Ground Lobby", "adult") == AccessibilityLevel.Normal) then
             return explo
           end
           if has("sticks") and explo ~= AccessibilityLevel.None then
@@ -11318,7 +11318,7 @@ data_per_region = {
           return has_explosives()
         end
       },
-      ["Gerudo Training Grounds Central Maze"] = {
+      ["Gerudo Training Ground Central Maze"] = {
         ["child_access"] = function()
           return AccessibilityLevel.Normal
         end,
@@ -11328,11 +11328,11 @@ data_per_region = {
       }
     }
   },
-  ["Gerudo Training Grounds Central Maze"] = {
-    ["scene"] = "Gerudo Training Grounds",
+  ["Gerudo Training Ground Central Maze"] = {
+    ["scene"] = "Gerudo Training Ground",
     ["dungeon"] = true,
     ["locations"] = {
-      ["Gerudo Training Grounds Hidden Ceiling Chest"] = {
+      ["Gerudo Training Ground Hidden Ceiling Chest"] = {
         ["child_access"] = function()
           local lot =
             (has("logic_lens_gtg") or (has("lens") and has("magic"))) and AccessibilityLevel.Normal or
@@ -11358,7 +11358,7 @@ data_per_region = {
           return and_accessibility(keys, lot)
         end
       },
-      ["Gerudo Training Grounds Maze Path First Chest"] = {
+      ["Gerudo Training Ground Maze Path First Chest"] = {
         ["child_access"] = function()
           if has("gtg_small_keys", 4) then
             return AccessibilityLevel.Normal
@@ -11376,7 +11376,7 @@ data_per_region = {
           return AccessibilityLevel.None
         end
       },
-      ["Gerudo Training Grounds Maze Path Second Chest"] = {
+      ["Gerudo Training Ground Maze Path Second Chest"] = {
         ["child_access"] = function()
           if has("gtg_small_keys", 6) then
             return AccessibilityLevel.Normal
@@ -11394,7 +11394,7 @@ data_per_region = {
           return AccessibilityLevel.None
         end
       },
-      ["Gerudo Training Grounds Maze Path Third Chest"] = {
+      ["Gerudo Training Ground Maze Path Third Chest"] = {
         ["child_access"] = function()
           if has("gtg_small_keys", 7) then
             return AccessibilityLevel.Normal
@@ -11412,7 +11412,7 @@ data_per_region = {
           return AccessibilityLevel.None
         end
       },
-      ["Gerudo Training Grounds Maze Path Final Chest"] = {
+      ["Gerudo Training Ground Maze Path Final Chest"] = {
         ["child_access"] = function()
           if has("gtg_small_keys", 9) then
             return AccessibilityLevel.Normal
@@ -11432,7 +11432,7 @@ data_per_region = {
       }
     },
     ["exits"] = {
-      ["Gerudo Training Grounds Central Maze Right"] = {
+      ["Gerudo Training Ground Central Maze Right"] = {
         ["child_access"] = function()
           if has("gtg_small_keys", 9) then
             return AccessibilityLevel.Normal
@@ -11452,11 +11452,11 @@ data_per_region = {
       }
     }
   },
-  ["Gerudo Training Grounds Central Maze Right"] = {
-    ["scene"] = "Gerudo Training Grounds",
+  ["Gerudo Training Ground Central Maze Right"] = {
+    ["scene"] = "Gerudo Training Ground",
     ["dungeon"] = true,
     ["locations"] = {
-      ["Gerudo Training Grounds Maze Right Chests"] = {
+      ["Gerudo Training Ground Maze Right Chests"] = {
         ["child_access"] = function()
           return AccessibilityLevel.Normal
         end,
@@ -11464,7 +11464,7 @@ data_per_region = {
           return AccessibilityLevel.Normal
         end
       },
-      ["Gerudo Training Grounds Freestanding Key"] = {
+      ["Gerudo Training Ground Freestanding Key"] = {
         ["child_access"] = function()
           return AccessibilityLevel.Normal
         end,
@@ -11474,7 +11474,7 @@ data_per_region = {
       }
     },
     ["exits"] = {
-      ["Gerudo Training Grounds Hammer Room"] = {
+      ["Gerudo Training Ground Hammer Room"] = {
         ["child_access"] = function()
           return AccessibilityLevel.None
         end,
@@ -11485,7 +11485,7 @@ data_per_region = {
           return AccessibilityLevel.None
         end
       },
-      ["Gerudo Training Grounds Lava Room"] = {
+      ["Gerudo Training Ground Lava Room"] = {
         ["child_access"] = function()
           return AccessibilityLevel.Normal
         end,
@@ -11495,11 +11495,11 @@ data_per_region = {
       }
     }
   },
-  ["Gerudo Training Grounds Lava Room"] = {
-    ["scene"] = "Gerudo Training Grounds",
+  ["Gerudo Training Ground Lava Room"] = {
+    ["scene"] = "Gerudo Training Ground",
     ["dungeon"] = true,
     ["locations"] = {
-      ["Gerudo Training Grounds Underwater Silver Rupee Chest"] = {
+      ["Gerudo Training Ground Underwater Silver Rupee Chest"] = {
         ["child_access"] = function()
           return AccessibilityLevel.None
         end,
@@ -11512,7 +11512,7 @@ data_per_region = {
       }
     },
     ["exits"] = {
-      ["Gerudo Training Grounds Central Maze Right"] = {
+      ["Gerudo Training Ground Central Maze Right"] = {
         ["child_access"] = function()
           return AccessibilityLevel.Normal
         end,
@@ -11523,7 +11523,7 @@ data_per_region = {
           return AccessibilityLevel.None
         end
       },
-      ["Gerudo Training Grounds Hammer Room"] = {
+      ["Gerudo Training Ground Hammer Room"] = {
         ["child_access"] = function()
           return AccessibilityLevel.None
         end,
@@ -11536,11 +11536,11 @@ data_per_region = {
       }
     }
   },
-  ["Gerudo Training Grounds Hammer Room"] = {
-    ["scene"] = "Gerudo Training Grounds",
+  ["Gerudo Training Ground Hammer Room"] = {
+    ["scene"] = "Gerudo Training Ground",
     ["dungeon"] = true,
     ["locations"] = {
-      ["Gerudo Training Grounds Hammer Room Clear Chest"] = {
+      ["Gerudo Training Ground Hammer Room Clear Chest"] = {
         ["child_access"] = function()
           return AccessibilityLevel.Normal
         end,
@@ -11548,7 +11548,7 @@ data_per_region = {
           return AccessibilityLevel.Normal
         end
       },
-      ["Gerudo Training Grounds Hammer Room Switch Chest"] = {
+      ["Gerudo Training Ground Hammer Room Switch Chest"] = {
         ["child_access"] = function()
           return AccessibilityLevel.None
         end,
@@ -11561,7 +11561,7 @@ data_per_region = {
       }
     },
     ["exits"] = {
-      ["Gerudo Training Grounds Eye Statue Lower"] = {
+      ["Gerudo Training Ground Eye Statue Lower"] = {
         ["child_access"] = function()
           return AccessibilityLevel.None
         end,
@@ -11572,7 +11572,7 @@ data_per_region = {
           return AccessibilityLevel.None
         end
       },
-      ["Gerudo Training Grounds Lava Room"] = {
+      ["Gerudo Training Ground Lava Room"] = {
         ["child_access"] = function()
           return AccessibilityLevel.Normal
         end,
@@ -11582,11 +11582,11 @@ data_per_region = {
       }
     }
   },
-  ["Gerudo Training Grounds Eye Statue Lower"] = {
-    ["scene"] = "Gerudo Training Grounds",
+  ["Gerudo Training Ground Eye Statue Lower"] = {
+    ["scene"] = "Gerudo Training Ground",
     ["dungeon"] = true,
     ["locations"] = {
-      ["Gerudo Training Grounds Eye Statue Chest"] = {
+      ["Gerudo Training Ground Eye Statue Chest"] = {
         ["child_access"] = function()
           return AccessibilityLevel.None
         end,
@@ -11599,7 +11599,7 @@ data_per_region = {
       }
     },
     ["exits"] = {
-      ["Gerudo Training Grounds Hammer Room"] = {
+      ["Gerudo Training Ground Hammer Room"] = {
         ["child_access"] = function()
           return AccessibilityLevel.Normal
         end,
@@ -11609,11 +11609,11 @@ data_per_region = {
       }
     }
   },
-  ["Gerudo Training Grounds Eye Statue Upper"] = {
-    ["scene"] = "Gerudo Training Grounds",
+  ["Gerudo Training Ground Eye Statue Upper"] = {
+    ["scene"] = "Gerudo Training Ground",
     ["dungeon"] = true,
     ["locations"] = {
-      ["Gerudo Training Grounds Near Scarecrow Chest"] = {
+      ["Gerudo Training Ground Near Scarecrow Chest"] = {
         ["child_access"] = function()
           return AccessibilityLevel.None
         end,
@@ -11626,7 +11626,7 @@ data_per_region = {
       }
     },
     ["exits"] = {
-      ["Gerudo Training Grounds Eye Statue Lower"] = {
+      ["Gerudo Training Ground Eye Statue Lower"] = {
         ["child_access"] = function()
           return AccessibilityLevel.Normal
         end,
@@ -11636,11 +11636,11 @@ data_per_region = {
       }
     }
   },
-  ["Gerudo Training Grounds Heavy Block Room"] = {
-    ["scene"] = "Gerudo Training Grounds",
+  ["Gerudo Training Ground Heavy Block Room"] = {
+    ["scene"] = "Gerudo Training Ground",
     ["dungeon"] = true,
     ["locations"] = {
-      ["Gerudo Training Grounds Before Heavy Block Chest"] = {
+      ["Gerudo Training Ground Before Heavy Block Chest"] = {
         ["child_access"] = function()
           return AccessibilityLevel.Normal
         end,
@@ -11650,7 +11650,7 @@ data_per_region = {
       }
     },
     ["exits"] = {
-      ["Gerudo Training Grounds Eye Statue Upper"] = {
+      ["Gerudo Training Ground Eye Statue Upper"] = {
         ["child_access"] = function()
           return AccessibilityLevel.None
         end,
@@ -11675,7 +11675,7 @@ data_per_region = {
           return and_accessibility(wall, climb)
         end
       },
-      ["Gerudo Training Grounds Like Like Room"] = {
+      ["Gerudo Training Ground Like Like Room"] = {
         ["child_access"] = function()
           return AccessibilityLevel.None
         end,
@@ -11705,11 +11705,11 @@ data_per_region = {
       }
     }
   },
-  ["Gerudo Training Grounds Like Like Room"] = {
-    ["scene"] = "Gerudo Training Grounds",
+  ["Gerudo Training Ground Like Like Room"] = {
+    ["scene"] = "Gerudo Training Ground",
     ["dungeon"] = true,
     ["locations"] = {
-      ["Gerudo Training Grounds Heavy Block Chests"] = {
+      ["Gerudo Training Ground Heavy Block Chests"] = {
         ["child_access"] = function()
           return AccessibilityLevel.Normal
         end,
@@ -11719,11 +11719,11 @@ data_per_region = {
       }
     }
   },
-  ["Gerudo Training Grounds MQ Lobby"] = {
-    ["scene"] = "Gerudo Training Grounds",
+  ["Gerudo Training Ground MQ Lobby"] = {
+    ["scene"] = "Gerudo Training Ground",
     ["dungeon"] = true,
     ["locations"] = {
-      ["Gerudo Training Grounds MQ Lobby Chests"] = {
+      ["Gerudo Training Ground MQ Lobby Chests"] = {
         ["child_access"] = function()
           return AccessibilityLevel.Normal
         end,
@@ -11731,7 +11731,7 @@ data_per_region = {
           return AccessibilityLevel.Normal
         end
       },
-      ["Gerudo Training Grounds MQ Hidden Ceiling Chest"] = {
+      ["Gerudo Training Ground MQ Hidden Ceiling Chest"] = {
         ["child_access"] = function()
           if has("logic_lens_gtg_mq") or (has("lens") and has("magic")) then
             return AccessibilityLevel.Normal
@@ -11745,7 +11745,7 @@ data_per_region = {
           return AccessibilityLevel.SequenceBreak
         end
       },
-      ["Gerudo Training Grounds MQ Maze Path First Chest"] = {
+      ["Gerudo Training Ground MQ Maze Path First Chest"] = {
         ["child_access"] = function()
           return AccessibilityLevel.Normal
         end,
@@ -11753,7 +11753,7 @@ data_per_region = {
           return AccessibilityLevel.Normal
         end
       },
-      ["Gerudo Training Grounds MQ Maze Path Second Chest"] = {
+      ["Gerudo Training Ground MQ Maze Path Second Chest"] = {
         ["child_access"] = function()
           return AccessibilityLevel.Normal
         end,
@@ -11761,7 +11761,7 @@ data_per_region = {
           return AccessibilityLevel.Normal
         end
       },
-      ["Gerudo Training Grounds MQ Maze Path Third Chest"] = {
+      ["Gerudo Training Ground MQ Maze Path Third Chest"] = {
         ["child_access"] = function()
           if has("gtg_small_keys", 1) then
             return AccessibilityLevel.Normal
@@ -11785,13 +11785,13 @@ data_per_region = {
           return AccessibilityLevel.Normal
         end
       },
-      ["Gerudo Training Grounds MQ Left Side"] = {
+      ["Gerudo Training Ground MQ Left Side"] = {
         ["child_access"] = function()
           if has("dinsfire") and has("magic") then
             return AccessibilityLevel.Normal
           end
           if has("bow") and has("firearrow") and has("magic") then
-            return access_region("Gerudo Training Grounds MQ Lobby", "adult")
+            return access_region("Gerudo Training Ground MQ Lobby", "adult")
           end
           return AccessibilityLevel.None
         end,
@@ -11799,13 +11799,13 @@ data_per_region = {
           return has_fire()
         end
       },
-      ["Gerudo Training Grounds MQ Right Side"] = {
+      ["Gerudo Training Ground MQ Right Side"] = {
         ["child_access"] = function()
           if has("sling") then
             return AccessibilityLevel.Normal
           end
           if has("bow") then
-            return access_region("Gerudo Training Grounds MQ Lobby", "adult")
+            return access_region("Gerudo Training Ground MQ Lobby", "adult")
           end
           return AccessibilityLevel.None
         end,
@@ -11814,18 +11814,18 @@ data_per_region = {
             return AccessibilityLevel.Normal
           end
           if has("sling") then
-            return access_region("Gerudo Training Grounds MQ Lobby", "child")
+            return access_region("Gerudo Training Ground MQ Lobby", "child")
           end
           return AccessibilityLevel.None
         end
       }
     }
   },
-  ["Gerudo Training Grounds MQ Right Side"] = {
-    ["scene"] = "Gerudo Training Grounds",
+  ["Gerudo Training Ground MQ Right Side"] = {
+    ["scene"] = "Gerudo Training Ground",
     ["dungeon"] = true,
     ["locations"] = {
-      ["Gerudo Training Grounds MQ Dinolfos Chest"] = {
+      ["Gerudo Training Ground MQ Dinolfos Chest"] = {
         ["child_access"] = function()
           if has("sticks") then
             return AccessibilityLevel.SequenceBreak
@@ -11838,7 +11838,7 @@ data_per_region = {
       }
     },
     ["exits"] = {
-      ["Gerudo Training Grounds MQ Underwater"] = {
+      ["Gerudo Training Ground MQ Underwater"] = {
         ["child_access"] = function()
           return AccessibilityLevel.None
         end,
@@ -11851,11 +11851,11 @@ data_per_region = {
       }
     }
   },
-  ["Gerudo Training Grounds MQ Underwater"] = {
-    ["scene"] = "Gerudo Training Grounds",
+  ["Gerudo Training Ground MQ Underwater"] = {
+    ["scene"] = "Gerudo Training Ground",
     ["dungeon"] = true,
     ["locations"] = {
-      ["Gerudo Training Grounds MQ Underwater Silver Rupee Chest"] = {
+      ["Gerudo Training Ground MQ Underwater Silver Rupee Chest"] = {
         ["child_access"] = function()
           return AccessibilityLevel.None
         end,
@@ -11868,11 +11868,11 @@ data_per_region = {
       }
     }
   },
-  ["Gerudo Training Grounds MQ Left Side"] = {
-    ["scene"] = "Gerudo Training Grounds",
+  ["Gerudo Training Ground MQ Left Side"] = {
+    ["scene"] = "Gerudo Training Ground",
     ["dungeon"] = true,
     ["locations"] = {
-      ["Gerudo Training Grounds MQ First Iron Knuckle Chest"] = {
+      ["Gerudo Training Ground MQ First Iron Knuckle Chest"] = {
         ["child_access"] = function()
           local weapon = AccessibilityLevel.None
           if has("sword1") then
@@ -11888,7 +11888,7 @@ data_per_region = {
       }
     },
     ["exits"] = {
-      ["Gerudo Training Grounds MQ Stalfos Room"] = {
+      ["Gerudo Training Ground MQ Stalfos Room"] = {
         ["child_access"] = function()
           if has("logic_gtg_mq_without_hookshot") then
             return AccessibilityLevel.Normal
@@ -11907,11 +11907,11 @@ data_per_region = {
       }
     }
   },
-  ["Gerudo Training Grounds MQ Stalfos Room"] = {
-    ["scene"] = "Gerudo Training Grounds",
+  ["Gerudo Training Ground MQ Stalfos Room"] = {
+    ["scene"] = "Gerudo Training Ground",
     ["dungeon"] = true,
     ["locations"] = {
-      ["Gerudo Training Grounds MQ Before Heavy Block Chest"] = {
+      ["Gerudo Training Ground MQ Before Heavy Block Chest"] = {
         ["child_access"] = function()
           if has("sticks") then
             return AccessibilityLevel.SequenceBreak
@@ -11922,7 +11922,7 @@ data_per_region = {
           return AccessibilityLevel.Normal
         end
       },
-      ["Gerudo Training Grounds MQ Heavy Block Chest"] = {
+      ["Gerudo Training Ground MQ Heavy Block Chest"] = {
         ["child_access"] = function()
           return AccessibilityLevel.None
         end,
@@ -11935,7 +11935,7 @@ data_per_region = {
       }
     },
     ["exits"] = {
-      ["Gerudo Training Grounds MQ Back Areas"] = {
+      ["Gerudo Training Ground MQ Back Areas"] = {
         ["child_access"] = function()
           return AccessibilityLevel.None
         end,
@@ -11959,11 +11959,11 @@ data_per_region = {
       }
     }
   },
-  ["Gerudo Training Grounds MQ Back Areas"] = {
-    ["scene"] = "Gerudo Training Grounds",
+  ["Gerudo Training Ground MQ Back Areas"] = {
+    ["scene"] = "Gerudo Training Ground",
     ["dungeon"] = true,
     ["locations"] = {
-      ["Gerudo Training Grounds MQ Eye Statue Chest"] = {
+      ["Gerudo Training Ground MQ Eye Statue Chest"] = {
         ["child_access"] = function()
           return AccessibilityLevel.None
         end,
@@ -11974,7 +11974,7 @@ data_per_region = {
           return AccessibilityLevel.None
         end
       },
-      ["Gerudo Training Grounds MQ Second Iron Knuckle Chest"] = {
+      ["Gerudo Training Ground MQ Second Iron Knuckle Chest"] = {
         ["child_access"] = function()
           return AccessibilityLevel.Normal
         end,
@@ -11982,7 +11982,7 @@ data_per_region = {
           return AccessibilityLevel.Normal
         end
       },
-      ["Gerudo Training Grounds MQ Flame Circle Chest"] = {
+      ["Gerudo Training Ground MQ Flame Circle Chest"] = {
         ["child_access"] = function()
           return has_explosives()
         end,
@@ -11995,7 +11995,7 @@ data_per_region = {
       }
     },
     ["exits"] = {
-      ["Gerudo Training Grounds MQ Central Maze Right"] = {
+      ["Gerudo Training Ground MQ Central Maze Right"] = {
         ["child_access"] = function()
           return AccessibilityLevel.None
         end,
@@ -12006,7 +12006,7 @@ data_per_region = {
           return AccessibilityLevel.None
         end
       },
-      ["Gerudo Training Grounds MQ Right Side"] = {
+      ["Gerudo Training Ground MQ Right Side"] = {
         ["child_access"] = function()
           return AccessibilityLevel.None
         end,
@@ -12019,11 +12019,11 @@ data_per_region = {
       }
     }
   },
-  ["Gerudo Training Grounds MQ Central Maze Right"] = {
-    ["scene"] = "Gerudo Training Grounds",
+  ["Gerudo Training Ground MQ Central Maze Right"] = {
+    ["scene"] = "Gerudo Training Ground",
     ["dungeon"] = true,
     ["locations"] = {
-      ["Gerudo Training Grounds MQ Maze Right Chests"] = {
+      ["Gerudo Training Ground MQ Maze Right Chests"] = {
         ["child_access"] = function()
           return AccessibilityLevel.Normal
         end,
@@ -12031,7 +12031,7 @@ data_per_region = {
           return AccessibilityLevel.Normal
         end
       },
-      ["Gerudo Training Grounds MQ Ice Arrows Chest"] = {
+      ["Gerudo Training Ground MQ Ice Arrows Chest"] = {
         ["child_access"] = function()
           if has("gtg_small_keys", 3) then
             return AccessibilityLevel.Normal
@@ -12047,7 +12047,7 @@ data_per_region = {
       }
     },
     ["exits"] = {
-      ["Gerudo Training Grounds MQ Underwater"] = {
+      ["Gerudo Training Ground MQ Underwater"] = {
         ["child_access"] = function()
           return AccessibilityLevel.None
         end,
@@ -12058,7 +12058,7 @@ data_per_region = {
           return AccessibilityLevel.None
         end
       },
-      ["Gerudo Training Grounds MQ Right Side"] = {
+      ["Gerudo Training Ground MQ Right Side"] = {
         ["child_access"] = function()
           return AccessibilityLevel.None
         end,