+++ ootrando_overworldmap_hamsda/changelog.md
@@ -7,6 +7,7 @@
 - added customizable setting preset loader
 - added settings for shuffled keys
 - changed hintable locations based on [PR #1381](https://github.com/TestRunnerSRL/OoT-Randomizer/pull/1381)
+- renamed locations based on [PR #1280](https://github.com/TestRunnerSRL/OoT-Randomizer/pull/1280) and [PR #1282](https://github.com/TestRunnerSRL/OoT-Randomizer/pull/1282)
 
 ## 3.6.0.0
 
@@ -885,7 +886,7 @@
   - tried to model the reqs
   - marked lens as optional (pretty easy to do the entire dungeon without)
   - added option to reach Bongo Bongo with longshot (there is a scarecrow you can reach)
-- improved Gerudo Training Grounds:
+- improved Gerudo Training Ground:
   - split up the items
   - added reqs to each of the outer chests
   - marked hammer as optional for the pillar room (can be easily skipped with a damage boost)